Renamed Feeder to Magazine; added initial commands.
