+++ src/main/java/frc/robot/subsystems/Magazine.java
@@ -7,27 +7,34 @@
 
 package frc.robot.subsystems;
 
-import static frc.robot.Constants.FeederConstants.topMasterPort;
 import static frc.robot.Constants.FeederConstants.bottomMasterPort;
+import static frc.robot.Constants.FeederConstants.topMasterPort;
 
 import com.ctre.phoenix.motorcontrol.can.WPI_TalonSRX;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 
-public class Feeder extends SubsystemBase {
-  private static Feeder instance;
+public class Magazine extends SubsystemBase {
+  private static Magazine instance;
   private WPI_TalonSRX topMaster;
   private WPI_TalonSRX bottomMaster;
 
-  public Feeder() {
+  /**
+   * Constructor.
+   */
+  public Magazine() {
     topMaster = new WPI_TalonSRX(topMasterPort);
     bottomMaster = new WPI_TalonSRX(bottomMasterPort);
     topMaster.setInverted(false);
     bottomMaster.setInverted(true);
   }
 
-  public static Feeder getInstance() {
+  /**
+   * Makes Magazine a singleton.
+   * @return Magazine instance
+   */
+  public static Magazine getInstance() {
     if (instance == null) {
-      instance = new Feeder();
+      instance = new Magazine();
     }
     return instance;
   }
@@ -37,7 +44,7 @@ public class Feeder extends SubsystemBase {
     // This method will be called once per scheduler run
   }
 
-  public void runFeeder(double speed) {
+  public void runMagazine(double speed) {
     topMaster.set(speed);
     bottomMaster.set(speed);
   }
